requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
Django==2.2.3
django-admin-tools==0.9.1
django-axes==5.9.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-constrainedfilefield==4.0.0
django-contrib-comments==1.9.2
django-crispy-forms==1.9.1
django-des==2.4.1
django-ipware==3.0.2
django-js-asset==1.2.2
django-model-utils==3.1.2
django-notifications-hq==1.6.0
django-phonenumber-field==5.0.0
django-session-timeout==0.1.0
django-solo==1.1.3
django-tinymce==3.0.2
django-tinymce4-lite==1.7.4
future==0.18.2
html5lib==1.1
httplib2==0.18.1
jsmin==2.2.2
jsonfield==3.1.0
oauth2client==4.1.3
phonenumbers==8.12.12

pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
reportlab==3.5.56
rsa==4.7
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
uritemplate==3.0.1
webencodings==0.5.1
whitenoise==5.2.0
